[dependency-injection.mjs] Add new module with items from type-driven-ui-basics.mjs
